+++ server/aws-lsp-codewhisperer/CHANGELOG.md
@@ -1,5 +1,77 @@
 # Changelog
 
+## [0.0.58](https://github.com/aws/language-servers/compare/lsp-codewhisperer/v0.0.57...lsp-codewhisperer/v0.0.58) (2025-06-23)
+
+
+### Features
+
+* surface file operation errors in tooltip ([#1713](https://github.com/aws/language-servers/issues/1713)) ([8d80e06](https://github.com/aws/language-servers/commit/8d80e06a18e89c1ae33430676ba461b2d7acd314))
+
+
+### Bug Fixes
+
+* **amazonq:** Handle throttling errors gracefully and give correct error message([#1733](https://github.com/aws/language-servers/issues/1733)) ([232e7ea](https://github.com/aws/language-servers/commit/232e7eac9556af3ab5e8cc86185b0c90b144cdd7))
+* fsReplace still available when agentic mode off ([#1731](https://github.com/aws/language-servers/issues/1731)) ([7904ea1](https://github.com/aws/language-servers/commit/7904ea18849bb5b9aa6c0e1eb4c6491f3d1598f4))
+* ide mapping for VS/Eclipse for send telemetry API ([#1724](https://github.com/aws/language-servers/issues/1724)) ([84373c5](https://github.com/aws/language-servers/commit/84373c537087492445dbf1d3c9d7b86254603ceb))
+* separate executeBash toolspec for mac and windows ([#1727](https://github.com/aws/language-servers/issues/1727)) ([33e0e4b](https://github.com/aws/language-servers/commit/33e0e4b2347e858ccb0c82c333aeaa8938b24c22))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aws/lsp-core bumped from ^0.0.9 to ^0.0.10
+
+## [0.0.57](https://github.com/aws/language-servers/compare/lsp-codewhisperer/v0.0.56...lsp-codewhisperer/v0.0.57) (2025-06-20)
+
+
+### Features
+
+* **amazonq:** bundle dependency events from workspace context server ([#1712](https://github.com/aws/language-servers/issues/1712)) ([587da41](https://github.com/aws/language-servers/commit/587da4152ed1273117fc549f49d0b81eef7d98a9))
+
+
+### Bug Fixes
+
+* **amazonq:** removed explanation field for mcp servers ([#1717](https://github.com/aws/language-servers/issues/1717)) ([cfc6831](https://github.com/aws/language-servers/commit/cfc683111307dc25c619177e0267860c096fcfe1))
+* make file collection for indexing non blocking ([#1701](https://github.com/aws/language-servers/issues/1701)) ([036efde](https://github.com/aws/language-servers/commit/036efdead9c68c4ee6e6590ee2e877ace4cabce6))
+* undefined path causing the loop to break ([#1718](https://github.com/aws/language-servers/issues/1718)) ([8e48b86](https://github.com/aws/language-servers/commit/8e48b866221c70c79156b714f036413816748b6c))
+
+## [0.0.56](https://github.com/aws/language-servers/compare/lsp-codewhisperer/v0.0.55...lsp-codewhisperer/v0.0.56) (2025-06-19)
+
+
+### Bug Fixes
+
+* **amazonq:** add ignore pattern for file events from workspace context server ([#1703](https://github.com/aws/language-servers/issues/1703)) ([7a0dd88](https://github.com/aws/language-servers/commit/7a0dd88a2f5251af8018084c55406cd8b9eaf51a))
+
+## [0.0.55](https://github.com/aws/language-servers/compare/lsp-codewhisperer/v0.0.54...lsp-codewhisperer/v0.0.55) (2025-06-19)
+
+
+### Bug Fixes
+
+* **amazonq:** serialize S3 uploads for file events from workspace context server ([#1700](https://github.com/aws/language-servers/issues/1700)) ([1884c57](https://github.com/aws/language-servers/commit/1884c5793d46227d871e8cf25c940f7a87795f04))
+
+## [0.0.54](https://github.com/aws/language-servers/compare/lsp-codewhisperer/v0.0.53...lsp-codewhisperer/v0.0.54) (2025-06-19)
+
+
+### Features
+
+* adding current working directory to the stdio transport for mcp… ([#1691](https://github.com/aws/language-servers/issues/1691)) ([02c4d64](https://github.com/aws/language-servers/commit/02c4d645a8c2778deab7af9f5377c26e99d01f20))
+
+
+### Bug Fixes
+
+* add fsReplace tool to batch edits ([#1533](https://github.com/aws/language-servers/issues/1533)) ([4125134](https://github.com/aws/language-servers/commit/4125134f6e7eee8276d6146a507834b3309c2ec5))
+* **amazonq:** itemid was accidentally removed by [#1689](https://github.com/aws/language-servers/issues/1689) ([#1698](https://github.com/aws/language-servers/issues/1698)) ([33524c0](https://github.com/aws/language-servers/commit/33524c092af8088705a3cbae09c6249ad5940ce6))
+* **amazonq:** profile is not set after re-auth ([#1690](https://github.com/aws/language-servers/issues/1690)) ([2a445ee](https://github.com/aws/language-servers/commit/2a445eef4cc2a70471fd1fc49e6ca4e301051442))
+* diff reports no lines added or removed ([#1549](https://github.com/aws/language-servers/issues/1549)) ([562f13e](https://github.com/aws/language-servers/commit/562f13e0223a8a01fefc9ca449aad02da9734709))
+* thinking doesn't get removed if response is empty ([#1699](https://github.com/aws/language-servers/issues/1699)) ([9a63c99](https://github.com/aws/language-servers/commit/9a63c99b3195c9da0f537980324998138f25a3fa))
+
+
+### Reverts
+
+* **amazonq:** bring back [#1684](https://github.com/aws/language-servers/issues/1684) ([#1697](https://github.com/aws/language-servers/issues/1697)) ([5e7aa76](https://github.com/aws/language-servers/commit/5e7aa76b6ebcf8e0a7489d3574cc14ed3d0ceebe))
+* **amazonq:** bring back [#1689](https://github.com/aws/language-servers/issues/1689) ([5b84b0e](https://github.com/aws/language-servers/commit/5b84b0e4c42c344d91ef9c99a04d3a2671221aa1))
+
 ## [0.0.53](https://github.com/aws/language-servers/compare/lsp-codewhisperer/v0.0.52...lsp-codewhisperer/v0.0.53) (2025-06-18)